+++ packages/expo-image/ios/ImageModule.swift
@@ -91,6 +91,18 @@ public final class ImageModule: Module {
         view.allowDownscaling = allowDownscaling ?? true
       }
 
+      Prop("autoplay") { (view, autoplay: Bool?) in
+        view.sdImageView.autoPlayAnimatedImage = autoplay ?? true
+      }
+
+      AsyncFunction("startAnimating") { (view: ImageView) in
+        view.sdImageView.startAnimating()
+      }
+
+      AsyncFunction("stopAnimating") { (view: ImageView) in
+        view.sdImageView.stopAnimating()
+      }
+
       OnViewDidUpdateProps { view in
         view.reload()
       }